% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acct.R
\name{AcctServer}
\alias{AcctServer}
\title{处理逻辑}
\usage{
AcctServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
AcctServer()
}
