% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workcenter.R
\name{btn_Update_WorkcenterServer}
\alias{btn_Update_WorkcenterServer}
\title{处理逻辑}
\usage{
btn_Update_WorkcenterServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
btn_Update_WorkcenterServer()
}
